Initial redesign of the Cash2 website
Reason for changes : to improve the looks ofthe Cash2 website. This is the initial commit. Still working on the CSS for responsive layouts. About 90% finished with the website redesign at this point. Still having some issue with getting tables to work in responsive layout, as well as with some margins and padding. Will continue to refine.